Updated SQL Functions
Added two functions which are connectors/updater of data within relational tables.
